requirements.txt
requests
pyyaml
keyring
cryptography